Node code is now much faster.
